The THaDetMap class and its child 'Module' structure have been upgraded to include first-logical-channel number information as well as the model number of the module. This removes the need for parallel 'fFirstChan' arrays. Also, a short list of model numbers (1872, 1875, 1877, 1881) has been included so a module can be tested to be an ADC or TDC with the 'THaDetMap::IsADC(Module *d)' and 'THaDetMap::IsTDC(Module *d)' methods.
This removes the detector-map-listing requirement to have an equal number
of ADC and TDC entries. The THaBPM, THaCherenkov, THaRaster and
THaScintillator classes have been modified to take advantage of these
features. The model-number is always OPTIONAL, and if not given the ADC/TDC
determination is done in the old way (ie: backward compatible).