Be certain that to be in the output file (fFile) during the THaOutput Init, keeping track of the previous state (copying gDirectory's value).
We had found a case where the user was reading in simulated data from
a ROOT file, and so gDirectory was not pointing at the correct directory
during initialization, and so the tree was not owned by the output file
but just languished in memory until analyzer->Close().

NOTE: the current directory is only important when CREATING a TTree or TH1;
you can be in any directory while ::Fill'ing it.