Make database keys case-sensitive again (=reverse last check-in).
This is the behavior I want for Release 1.5 and up. Release 1.4 folded the
case for the part of the key that was given to LoadDBvalue() - usually just
the last part of the key excluding the prefix. This was confusing -
case-sensitivity is what ROOT users expect because that's how all user strings,
object and class names, identifiers etc. are treated.

This change may require minor modifications to some user databases, but
it will be immediately obvious from error messages which entries require
changes.