Make search for matrix element section header [ <prefix>.global ] case-sensitive, consistent with the rest of the database system.
In THaVDC, print warning if a particular configuration is requested
(fConfig != "") but not found, and load the first found section as default.
Also, load the first section even if it starts with a config header and
no particular configuration was requested. (Previously, if fConfig == "",
nothing would get loaded if the first section started with a tag.)
Example in db_L.vdc.dat:

[ L.global ]
some numbers that are now completely ignored no matter what the comment says
[ config=highmom ]
comment line describing the matrix elements, ignored
t 0 0 0  -1.000 0.05 0.0002 0.000 etc...

This section will now be loaded if fConfig == "" or fConfig == "highmom".