Fixed cluster association bug where only the earliest indexed clusters would be associated.
Included enforcement of "V shape" time structure for clusters.  This
should have additional checks on the step size enforced.

Changed code to consider all TDC hits, not just the earliest.

Tracks will not be produced if there is more than one cluster on any
plane in the specified time window, as was originally intended.