Add support for multi-value text variables in database files.
Multi-valued variables probably only make sense on the LHS. For instance

L.${alldet}.debug = 1

expands to multiple keys, one for each detector, which can be useful
in certain cases. (This particular case can, of course, also be handled
with a search argument in the DBRequest.) However

L.debug = ${allvalues}

is basically useless. If ${allvalues} = (1,2,3) then this line is just an
ineffective way of writing L.debug = 3.