Rename various VDC classes and variables for clarity
Specifically, "UVTrack" is renamed to "Point" and "TrackPair" to
"PointPair". A "point" is a pair of one U and one V cluster in the
same UV plane (chamber). The U and V coordinates can be combined
to make a "local track" (x,y,x',y'), which we now call a "point".
A "point pair" is a tentative association between one point in the lower
chamber and one in the upper chamber.