Clean up VDC spacing parameters and implement THaVDCPointPair::Print
The VDC class now keeps only one spacing parameter, fSpacing, which
is identical to the old fUSpacing, the spacing between the reference planes
of the upper and lower chambers.

THaVDCPointPair must be given the chamber spacing when being constructed.
(This is for efficiency; in principle, the spacing can be derived from the
upper and lower points.)

Move the printing of debug information for a point pair from
THaVDC::ConstructTracks into THaVDCPointPair::Print.