Add support for continuing database lines without continuation characters
Any lines not containing an '=' character are automatically considered
continuation lines until the next blank line or next line with a '='
is reached. Comment lines within continuation blocks are allowed.
The style using continuation characters, "\", and this new style must
not be mixed within a given block, but different styles may be used in
different blocks of a file.

See the updated database files in DB/20121201 for examples.