Add default method THaAnalysisObject::MakePrefix()
This sets the prefix to "name.", which is a reasonable default for
many modules, e.g. physics modules and apparatuses.

As a result, THaAnalysisObject is no longer an abstract base class.
However, since the constructor is protected, we still prevent people
from instantiating plain THaAnalysisObjects. This is intentional
as such an object would have no functionality.