Support specific elements of vararray variables in THaFormula
It is now possible to define formulas containing an expression like
"var[n]", where var is a variable-size array global variable and
n is any integer, regardless of the actual size of the array at
definition time. If n >= the size of the array at evaluation time,
var[n] evaluates to kBig. That's probably not quite yet what we want.
We actually want the whole formula to evaluate to kBig and/or report
something like IsValid() == false.

Putting just "var" in an expression is currently equivalent to "var[0]".
This leads to the issue of multi-valued formulas which needs additional
though.