Implement THaCut array evaluation modes (AND, OR, XOR)
If the cut expression is an array, the cut expression is evaluated
for each array element. If the array is empty (variable-size array),
the cut evaluates to false. Otherwise, the cut result is the
AND, OR, or XOR of the tests of the individual elements. Here, XOR
means that the expression must be true for exactly one element.
Aliases for the modes are ALL, ANY, and ONE/ONEOF.

To specify the mode, prefix the cut expression with the mode and a
colon, for example

THaCut* cut1 = new THaCut("cut1", "OR:xarr>0", "DEFAULT");

If no prefix is given, the mode defaults to AND.

Resolves #58