Add support for building a bundled copy of EVIO to the Make system
The idea is that analyzer distribution download will have an EVIO tarball
bundled. If EVIO_LIBDIR or EVIO_INCDIR are unset at build time, this
tarball will be compiled. If no bundled tarball is present, the Makefile
will attempt to download it from coda.jlab.org. This would be the case,
for example, when trying to compile a clone of the git repository without
having EVIO installed. All this should make things more or less
foolproof even for non-expert collaborators. (We hope.)