Remove obsolete scaler code and NormAna library
The old scaler code is superseded by the event type handlers for scalers,
which are different for Hall A and C.

NormAna really an SDK example that must be customized per experiment.
Also it relied on the now-defunct THaScaler and needs to be modified
to use the new scaler event type handlers.