Fix incorrect calculation of error correlation coefficient in VDC cluster fit
This affected the calculation of the fit intercept error, which translates
into the cross-over position error. Fortunately, the value was not used any
further in the tracking algorithm (though it probably should be), only
exported as a global variable.

This bug was already found and fixed in the upstream advanced reconstruction
code (on 9-Jun-2014).