Allow clean re-Init() of Fadc250Module. Compile debug code conditionally
Debug code is only compiled if WITH_DEBUG is defined, consistent with
the main code. If one wants the class always to write a debug file
automatically, uncomment #define DEBUG in the code and recompile;
otherwise call SetDebugFile externally.

Re-init is a bit tricky. Note that someone may set fDebugFile externally
via Module::SetDebugFile, so we cannot assume that we own this object.
The SetDebugFile logic is not very clean. It would be better to have
a method like SetDebugFileName and let the class handle the file IO
implementation internally, so the Module would always own the
fDebugFile object.