Add basic implementation of dbconvert utility for old->new database format
Comments:
The behavior of CopyFile::ReadDB mostly matches that of LoadDBvalue.
There is one difference however: The LoadDB logic allows keys to be
undefined in certain periods when using subdirectories. dbconvert
doesn't support validity end dates so far, and so the last key valid
before a directory start time will always continue to be valid until
the first timestamp in a file in that directory. This is the intended
behavior; the behavior of the old LoadDB code is flawed and perhaps
should be fixed in the future.