Various dbconvert bugfixes & improvements
- optimize result of InsertDefaultFiles
- small code improvement in PruneMap
- suppport TriggerTime and BigBite shower databases
- fix logic of default values for certain keys

Since we are considering the entire database, we need to explicitly
create all keys first for all times, even if they are equal to the
default value (either because redundantly set to the default or
because undefined) for some of the times. Once all input files
for a given detector have been parsed, we can then check if certain
keys always equal the default. Only if so, we can omit them in the
output. If we did otherwise, we would have non-default values carrying
over into time ranges where defaults apply.