From Bob Michaels: Add bank decoding support to decoder
These changes support bank decoding.

If the crate map has lines like this for a FADC250 module in crate 10,
slot 4, bank 3, then that module will receive the data. A "bank" is
well-defined in CODA. Most newer setups tend to put the data of one
module into its own bank.

==== Crate 10 type vme
[pound sign] slot model bank
4 250 3

This is in contrast to the "header decoding" that was previously the
only choice.

The "all histogram identical" tests show that the code did not break
the old code.