Don't depend on LD_LIBRARY_PATH while building!
Because libdc.so was linked to libevio.so, libevio.so had to be in
the LD_LIBRARY_PATH when linking the main analyzer executable, regardless
of EVIO_LIBDIR. Oops. Of course, we shouldn't link libdc.so to libevio.so.
So now we explicitly link the main analyzer to libevio.so (with EVIO_LIBDIR
as the search path) like all the other shared libs.