Remove scintillators and "trigger time" detector from default HRS setup
Although the HRS always has scintillators, they are not always named
"s1" and "s2", and the number of scintillator planes (=detectors) may
vary. So we now leave it up to the user to configure these.

We might print a warning if Init detects that there are no scintillators
in the HRS at all. Then again, maybe not - someone might simply want
to do VDC debugging without bothering with scintillators.

The trigger time detector has never been active since aournd 2005 as far
as I can tell. It has a hardcoded name "trg", for which I cannot find
any database files on the adaq cluster - anywhere. We should not
set up detectors that are only used by experts unless they have a
stable default configuration.