Fix silly regressions in Fadc250Module introduced by my recent updates
I had forgotten to update the signature of PopulateDataVector.
While I'm at it, let's fix SumVectorElements as well (which really was
inefficient already in the original).

Also, disable creation of a separate debug file in Fadc250Module.
For some reason to be determined, overriding the debug file stream causes
nasty crashes. The whole design of manually allocating an ofstream on the
heap for debug message logging is really dorky anyway. Someone needs to
clean up that mess.