Add multiblock capability to decoder
To add pipelining capability, reorganize so that 3-arg and 4-arg version of
LoadSlot are tied together, not sure we need both.  load THaSlotData
objects with one private method instead of two blocks of code.
renamed Decode as DecodeOneWord because of the argument it uses, it
was a private method so it does not matter.  The public interface
remains the same but is augmented to support pipelining and to support
THaEvData::GetData methods

Add more EModuleTypes

Add block header to each event since FADC250 decoding uses block header
to get slot number

Add fFirmwareVers variable; some behavior is firmware dependent.