Check if slot is used.
Added a check to the `CodaDecoder::bank_decode` which checks if slot is
used before getting the slot number. The problem is that the method
loops over all slots from `minslot` to `maxslot`, which is fine if all
slots are filled in the cratemap. If however one uses slot 8 and 10, but
not slot 9, this loop will get a `bank` default value of -1, which
doesn't pass further tests. Now a slot that is not used is just skipped.