dbconvert: replace popt with getopt_long handle program options
popt is not as readily available as I had thought, and it doesn't
offer a really compelling advantage here (other than less work for me).

In the process of this replacement, I wrote a couple of fairly generic
supports classes/functions for working with program options:

- wrapping_ostream writes text wrapped at a certain width
  (default = terminal window width)
- print_usage() and print_help() pretty-print usage and help info

These could be moved into a separate source file at some point.