VDC algo change #1: Don't make tracks if multiple clusters in both U an V
It is rare, except at very high rates, to get true multi-track events
in the VDCs. With the present VDC reconstruction algorithm, these
events simply cannot be reconstructed reliably. In fact, if both
U and V planes of a chamber have multiple hits, the present algorithm
would corrupt good tracks in at least 50% of the cases. Since this
case is obviously not supported by this code, it is better not to
generate any tracks rather than potentially incorrect ones. Users can
easily count the frequency of multi-cluster events with tests on the
number of clusters per plane.