VDC algo change #2: Don't reuse UV track combinations that share clusters
This was definitely a bug in the existing code, causing excessive
ghost tracks. If a combination of U and V clusters in a chamber
("THaVDCUVTrack") has been assigned to a reconstructed track,
neither of these U and V clusters should be reused by any other
reconstructed tracks. Doing so almost certainly would result in
a ghost track with poorer reconstruction quality.