Add the BPM x/y coordinates from fOrigin to the calculated beam position
This maintains consistency. The "offset" from the database is intended
to be a (presumably small) calibration/alignment correction to the
nominal, surveyed BPM position that fOrigin is set to.

If you want to put any sort of non-zero x/y shift of the BPM position
into fOffset, just make sure that you set x and y of fOrigin (the
"position" database parameter) to zero.