Add THaDecData::Reset method, useful for testing in standalone scripts
THaDecData does *not* delete previous channel definitions when
re-initializing. (This behavior was inherited from the original
implementation.) The reasons is that each channel definition also
creates a global variable, and global variables are not supposed
to disappear in the middle of an analysis.

To allow for a complete re-initialization, we've added Reset.
Sometimes one just wants to clear out all definitions and load
a new configuration from a different database. However,
Reset should not be called if the global variables from the channel
definitions are being used elsewhere, at least not without then
reinitializing all the THaDecData-related global variables.