SCons: Remove unneeded root dictionary intermediaries
Instead of generating first the dictionary .C source via rootcint,
then explicitly an .os shared object via "SharedLibrary", which is
then one of the "sources" of the respective application shared library
(libHallA, libdc), simply specify the dictionary .C file as a source
for the application library. rootcint generates the .C file, and SCons
knows how to compile it for inclusion in the application library.