Address "unused parameter" warnings, mostly in decoder classes
Most of these warnings seem completely harmless, but in a few places
it looked like something was forgotten or incomplete. In those cases,
I left the warning, added a TODO or actually used the parameter.

The Decoder::Module::Decode method might need to be revisited.
There are several modules which implement only a dummy Decode function.
Could that become the default behavior then?