Makefile: Add more options and safer detection of pre-installed EVIO
EVIO is searched for in the following order:

1. $EVIO_LIBDIR and $EVIO_INCDIR (both must be defined)
2. $EVIO (look in subdirs $(EVIO_ARCH)/{lib,include})
3. $CODA (look in subdirs $(EVIO_ARCH)/{lib,include})

In each case, the Makefile checks if the include directory contains
the file "evio.h" and the library  a file "libevio.*". If both exist,
the corresponding locations are used for building and linking the analyzer.

If none of these locations appear to contain a valid EVIO installation,
the source is downloaded, and EVIO is built locally.