dbconvert: replace deprecated readdir_r with readdir
Why were we using the reentrant version anyway?

Better yet would be to use the platform-independent ROOT API to access
the file system. But all of dbconvert is written with reliance on
Unix-specific library calls, so that would entail a large amount of
refactoring. Probably not worth it, given most users will rarely use
this utility.