Add optional plugin library with v1.5 VDC tracking code
This plugin can be used, for example, for comparison studies or to
reproduce old results. The code in the library is identical to the
that of version 1.5.37 except that the database functions have been
converted to read the new v1.6 format. Thus, there is no need to
maintain separate VDC databases for the old and new VDC algorithms.

After testing it, I should probably also make the improved
v1.5.38 VDC code available here, where the improvements can be
enabled/disabled via a flag.