Re-enable reading of configurable VDC v1.5 parameters via database
Reading of "tdc.min", "tdc.max", "tdc.res" and "maxgap" was commented
out in the OldVDCPlane database reader. Re-enable these so that
if they are set in the new-style VDC database, they are read from there.

In this way, parameters supplied to the old and new VDC code stay
consistent, and OldVDC can be used to study /algorithm/ differences,
not differences in database parameters that accidentally creep in because
of technical differences (database vs. Set functiions).

These parameters can also be set via Set functions, in which case the
value supplied explicitly overrides the database value. We use
OldVDCPlane::fBits 14-17 as flags to indicate which Set function was
called. This is supported only for backward-compatibility with old
scripts.

Closes Redmine issue 250.