Set sign of correlation bewteen time and raw TDC data based on TDC mode
Add TDC resolution parameter to THaCherenkov.

Historically, a negative TDC resolution has been set in the scintillator
database to indicate common stop TDC mode. The new code tries to be
backward compatible: If the TDC mode is not explicitly set (old database),
the sign of the TDC resolution is kept. If the mode is set, the TDC
resolution is taken as positive, and the sign is applied based on
the TDC mode.