Handle conflicting VarDef specs in THaVarList::DefineVariables
The original version would not detect a situation where size>1,
indicating a fixed-size array, and count!=0, indicating a
variable-size array. Admittedly, that's a user error, but the code
should still deal with it. The result was an ill-formed variable-size
array variable with an array name like "a[10]".

The fixed version prints a warning if size>0 and count!=0 and
assumes that the user wants to create a variable-size array.
Variable-size arrays should always be specified with size==0.

Additionally, the new version allows definition of std::vector
variables by giving a type like kIntV etc.

Closes #153