Fix from Bob Michaels for broken F1TDC bank decoding
Comment from Bob: This fixes the problem Alexandre was noticing that
bank-decoding of F1TDCModule did not work. I'm not sure how it ever
worked but maybe it was previously used in some test code that called
the 3-arg version of LoadSlot explicitly.

After discussion with Bob, we felt it would be fine to put the
fix into the base class, Decoder::Module, as the default implementation
of the 4-parameter version of LoadSlot. In this way, this problem should
be unlikely to come up again for any other decoder module as well.

Fixes Redmine bug # 349.