Split project into two separate libraries: Podd and HallA
libPodd holds the general (framework) classes, while libHallA contains
classes specific to Hall A replay and analysis.

The current split is at the level of source files. There are a few instances
where specific Hall A code is part of a normally more general class. This
still needs to be sorted out.

Also
 - move CMake helper scripts to cmake/scripts subdirectory
 - move various example/utility scripts to more appropriate locations
 - move main programs to apps directory. Rename main.cxx to analyzer.cxx
 - move OldVDC and VDCsim into new plugins directory