Refactor THaDecData into general and Hall A-specific parts
The general class is Podd::DecData, while the Hall A class
keeps the name THaDecData for backward compatibility.

Because of the way the plugin registration system works, once
libHallA is loaded, even Podd::DecData will recognize the
Hall A-specific TrigBitLoc decoder. It will correctly initialize
and process channels defined as this type, except that the
trigger bit pattern variable "evtypebits" will not be filled.
"evtypebits" will only be filled with TrigBitLoc data by THaDecData.