Make decoder package independent of main analyzer code
There was a circular dependency between the decoder and the analyzer
libraries which required building both libraries to be linked allowing
undefined symbols. This commit removes that circular dependency.
The decoder library now always compiles in "STANDALONE" mode (except
that the decoder test executables are not built without setting
STANDALONE).

Note that the analyzer must now use Podd::CodaRawDecoder instead of
Decoder::CodaDecoder for the crate map database file to be found in the
usual directories. Decoder::CodaDecoder only searches in the current
working directory now. THaInterface sets the default decoder class
accordingly.