Fix build of decoder standalone executables
These executables acutally do not need to be linked directly to EVIO,
and none of the programs need to include evio.h. All low-level EVIO
interactions occur through the linked decoder library, just like
with the main analyzer.

The SCons build system could be organized a little better so that
it wouldn't be necessary to FindROOT again in hana_decode/apps.
To build the standalones, add standalone=1 to the scons command line.

The CMake system gained an explicit STANDALONE configration option
(apparently forgotten earlier). To enable the standalones, add
-DSTANALONE=1 to the cmake command line. (Previously you needed
-DSTANALONE:BOOL=1.)

Also, removed a few unnecessary #includes from Fadc250Module.
The evio.h include in the header was a recent regression. Weird.