Better error checking in standalone progs. Disable non-functional parts
With Podd 1.6+, EPICS and scaler handling does not work in the way it
was originally implemented in the example programs. As a result, the
relevant example programs crash on assertions indicating usage of
unimplemented functions (e.g. "GetScaler"), exactly as it should be.

We should update those programs to use the new event handler
mechanism for EPICS and scaler data.