New header DataType.h
Make the type Data_t (typically Float_t or Double_t), which was
introduced in the update of THaScintillator, available for other
classes as well. Data_t can be used for event-by-event data and
conditions.

DataType.h is also a good place for a central definition of kBig,
which we use throughout the code to indicate invalid/unset data.

Let data type be Double_t - we have enough RAM these days