Refactor Decode() methods. Add detectors with FADC decoders.
This takes advantage of the hit iterators introduced in THaDetMap.

Also, refactor & clean up DefineVariables methods.
Each AnalysisObject's DefineVariables method is now called from
DefineVariablesWrapper, which takes care of setting and clearing the
status flag fIsSetup.

fIsSetup should no longer be tested or modified in DefineVariables,
although doing so will not normally break anything (it'll just be
messier than it needs to be), so this is backwards-compatible.

Like THaAnalysisObject, the DetectorData classes now call the actual
virtual method that defines their variables through a wrapper that
manages the initialization status flag (fVarOK).

Since there is no Init method here, it seems a like a good idea to
keep the name DefineVariables for the outer wrapper function. This is
the method that's visible to clients. The virtual method, which each
DetectorData object should implement, is DefineVariablesImpl.

As with THaAnalysisObject, there is no need anymore to test or modify
fVarOK in DefineVariablesImpl. It should be set correctly if variables
are set up and removed through DetectorData::DefineVariables.