THaSlotData: Refactor with std::vector arrays. Don't limit size of data[]
There is no compelling reason to have an arbitrary hardcoded limit on the
number of data words per channel. (The previous value was 128k.) If memory
is available, which it usually is, then it should be used.

The std::vectors for the event data use a custom allocator that leaves
memory uninitialized on resize() for better performance. This is consistent
with the behavior of the C-arrays in the previous version.

Possible improvements:
- Keep memory usage statistics
- Log a warning if memory usage does grow unusually large
- Shrink memory footprint again if prior growth appears due to an outlier
- Catch bad_alloc exceptions to either terminate or skip the event