CodaDecoder: Delay crate map and crateslot init until first physics event
The crate map and the crateslot array are only needed for physics analysis,
so unless we hit a physics event, there is no point in initializing them.

The previous version caused the crate map and crateslot array to be
initialized 4 times: First, during the THaRun->Init)() phase,
once for the default run time, then for the run time retrieved from the
prestart event. These two initializations were then repeated for the
main analysis phase.