Add support FADC decoding to BPM, Raster and RasteredBeam
Includes the following commits:

- THaBPM: Don't count hits with out-of-range logical channel number

- Add FadcRaster and FadcRasteredBeam classes

- THaDetMap: Allow ignoring logical channel in detmap tuple

This is useful in the THaRaster class where the detector map includes the
logical channel but effectively ignores it (due to some bizarre
misunderstanding of things, I guess). To activate this feature, replace
kFillLogicalChannel with kSkipLogicalChannel in the flags for the call to
FillDetMap. As a result, the detector map will be read as before, but the
logical channel number will increment continuously in the order in which
modules are defined, ignoring any number given for the "first" channel.

- THaRaster: make use of new Decode() scheme from THaDetectorBase

- THaDetMap: Save actual module type in hitinfo

This is a copy of THaDetMap::Module::type, not just a simple ADC/TDC flag.

- THaDetectorBase: Don't print multihit warnings for multifunction modules

Multifunction modules typically report multiple "hit", where the hit number
is actually a word index. If the user chooses the single-hit iterator for
such a module, assume they are interested in the first word only (usually
the main measured quantity like amplitude or time).

- THaRaster/THaBPM: Use Int_t for NCHAN etc. for fewer conversions