CodaDecoder: Fix off-by-one error of ROC length in FindRocsCoda3
To stay consistent with the meaning of the variables in RocDat_t,
the length needs to be the length of data following the length word,
not the total length of the ROC chunk. In short:
pos+len = position of last word of data

Conversely, any routines that take a start and end pointer into the
event buffer assume that the end pointer (pstop) points to the last word
of data. So, pstop = pstart + len-1.

Also, the meaning of pos and len in BankDat_t is yet different again. Here,
len IS the total length of the bank data, including the start word at pos,
so pos+len = one word past end of data.

These are simply the conventions adopted in the decoder from the CODA2
days. They are not necessarily the best but at least so far as the API is
concerned, we should stick to them or derived code will break.

There were also related subtle off-by-one errors in "end of data" checks
in some decoder modules, perhaps because the respective authors at various
times had a different idea of the meaning of len.

Updated comments in various places to help clarify this.