Database: Fix buggy commit d7671ff6 (time zone offsets in timestamps)
On macOS (Catalina at least), strptime with the "%z" time zone
conversion specifier sets errno to a spurious value (84, "Value too
large") even if succeeding otherwise. This may be a system library bug.

On Linux, strptime parses and converts the "%z" specifier, but does not
convert the resulting struct tm to localtime, as macOS does. Hence, it
is necessary to convert the result explicitly to localtime.

Also, the code in IsDBdate for conversion to localtime didn't work as
intended.