THaAnalyzer: Refactor PrintSummary code. Write more info to summary file
There are now several virtual functions that allow derived analyzer
classes to implement desired changes to how the summary is written to
both stdout and file (if configured).

The summary file is always written to in append mode by the predefined
functions. If fOverwrite is true (default), the summary file is
truncated at the start of the analysis. We assume the user only writes
one summary per analysis job, even if analyzing multiple run files. If
the summary file name is changed between calls to Process(), truncating
of files other than the first will not occur.