CodaDecoder: Let event_num always hold last physics event number
Resetting event_num at the start of LoadEvent probably messed up
event numbers stored with all EPICS data as well as invalidated
usages in any other non-physics event handlers.

To do: We should add a way to reset the event number somehow.
Currently, for CODA3 at least, the number just keeps counting for
the lifetime of the decoder object. This could potentially be
an issue when analyzing multiple runs in the same session, for
example.