CMake: Use git commit date, not CMake run date, for HA_DATE and DATETIME
Also, print out HA_DATE as part of the analyzer version string.

If git is not available, or the source is not a git repository (e.g.
tarball distribution), quietly revert the old behavior of using the CMake
configuration time as the analyzer "date".

The ha_compiledata.h header should be automatically re-generated whenever
the git status changes. The logic was adapted from the CMake scripts
discussed in https://stackoverflow.com/questions/1435953/