Also decode user event type 137 (additional DAQ configuration info)
The decoded strings are appended to the vector of strings obtained from
event type 138, which are stored in DAQconfig both in the decoder and in
the current THaRun object.

The DAQconfig structure in the decoder is cleared for each new run. This
should do the right thing if multiple runs are replayed in a single
session and each run is represented by its own THaRun object.