CMake: Allow reuse of PoddCompileInfo.cmake by parent projects like hcana
This renames the generated git_description.h to git_description_Podd.h.
This is fully transparent so long as a dependent project includes only
ha_compileinfo.h, as documented.

Reuse of the PoddCompileInfo module by a parent project generates
${CMAKE_CURRENT_BINARY_DIR}/git_description_${PROJECT_NAME}.h
which the project can use as it sees fit. For example, hcana includes
it in hc_compileinfo.h.