CodaDecoder: Bugfix: Use CODA3 event number from DAQ
To match the CODA2 behavior, for CODA3 data we should set the event
number to the value that the DAQ stores in the data file, not to that
of an internal counter that counts how often physics_decode has been
called.

The CODA3 event number is encoded in the trigger bank at the start
of every physics event. We already decode this bank, so this is trivial
to implement.