THaPrintOption: Add Contains(), ToLower(), ToUpper(), operator<<
Also, ensure clean conversion to values. Previously, an option like
"123opt" would be assigned a value of 123, which is nice but probably
not what was intended. With this change, the assigned value will be 0,
as it is for text options. Only numbers-only options will be converted
to non-zero values.

FWIW, also add move constructor and assignment operator. Make
constructors non-explicit so copy initialization works:

THaPrintOption opt = "HELLO,WORLD"