CodaDecoder: Ensure consistent skipping of banks not found in event
If a bank was defined in the crate map but not found in the current
event, bank_decode() would skip any modules associated with this bank,
as it should. However, LoadFromMultiBlock would try to process them,
effectively making calls on incorrectly initialized modules. This is
fixed now. It only affects multi-block mode, of course.