CMake config file: Save PACKAGE_PREFIX_DIR across find_dependency call
A dependency config file may overwrite PACKAGE_PREFIX_DIR, causing the
subsequent include of the Podd targets file to fail. This is true in
particular if the dependency generated its project config file with
CMake's configure_package_config_file() command.

This is currently not an issue with Podd's ROOT dependency, so we're
doing a bit of defensive programming here. It IS and issue with hcana
though (see hcana commit 6bf09d67, to be updated).