Make prescale factors signed int everywhere and store in std::array
For historical reasons, the prescale factors haven been signed integers.
Changing the type to unsigned, as it should be, could lead to unexpected
issues, so I'll leave it alone. This commit eliminates the type mismatch
or the prescale factors in CodaDecoder and THaRunParameters. Both are
signed integer now. Also, there is little point in using std::vector
to store the prescales since the array size will never vary unless there
is a major change in the DAQ, so switch to std::array.

While I was at it, I did some light refactoring of the prescale_decode
routines in CodaDecoder (fewer lines of code, additional error checks).

TODO: TS_PRESCALE_EVTYPE does not yet seem to be supported with CODA 3.