CodaData: Make "unexpected end of file" a fatal error, as it should be
Here, "fatal error" means an error that gracefully exits the analysis,
properly closing the output file, saving the summary, etc., as if a
regular EOF had been encountered.

Also, let analyzer print counters and cut summary even if the run
terminated with a fatal error. Fatal errors seem to occur most often due
to truncated EVIO files, so why not report the statistics of the
successfully completed analysis up to the truncation point.